backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.118.0
fastapi-cli==0.0.13
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pip==25.2
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.1.0
rich-toolkit==0.15.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.48.0
typer==0.19.2
typer-slim==0.19.2
typing_extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
